refactor: Login, Register 컨트롤러 추상화
